+++ history.md
@@ -40,10 +40,14 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 
 ## List of versions
 
-## version 0.9.0-beta.0 - _(Unreleased)_ - 2026-?-? {#0.9.0-beta.0}
+## version 0.9.0-beta.1 - _(Unreleased)_ - 2026-?-? {#0.9.0-beta.1}
 
 - [x] (Added) _App_ Align menu's in Windows with macOS (PR #3243)
-- [x] (Added) _App_ Swift-based MacOS desktop app instead of Electron (PR #3238)
+- [x] (Fixed) _App_ prevent accidental navigation when the backspace key (PR #3251)
+
+## version 0.9.0-beta.0 - 2026-08-31 {#0.9.0-beta.0}
+
+- [x] (Added) _App_ Swift-based macOS desktop app instead of Electron (PR #3238)
 - [x] (Added) _App_ WPF-based Windows desktop app instead of Electron (PR #3234)
 
 ## version 0.8.2 - 2026-08-27 {#v0.8.2}